Refactored active-state
