+++ js/map-layer.js
@@ -1,0 +1,11 @@
+/* global L:readonly */
+
+const tileLayer = L.tileLayer(
+  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+  {
+    attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+  },
+);
+
+
+export {tileLayer};
